Implement ApplyChanges for the Azure DNS provider.
This commit implements the ApplyChanges method for the Azure DNS provider.

The provider maps requested changes to the most specific Azure DNS zone which
should host the DNS records.  If a suitable zone is not found, the change request is
ignored.

The provider will create A records for the domains it manages; when using the
TXT registry, a corresponding TXT entry will be created for each managed
domain.